save progress

+++ PartB_RUDP/RUDP_API.h
@@ -24,14 +24,14 @@
  * @brief Creates an RUDP socket and a handshake between two peers.
  * @param struct sockaddr_in* and the peer type (CLIENT or SERVER);
 */
-int rudp_socket(struct sockaddr_in *my_addr, int peer_type, uint8_t *seq_number);
+int rudp_socket(struct sockaddr_in *my_addr, int peer_type, uint16_t *seq_number);
 
 /* 
  * @brief Sending data to the peer. Waits for an acknowledgement package, if not received, resends the data.
  * @param 
  * @return 
 */
-int rudp_send(int sock_id, void *data, size_t data_size, int flags, struct sockaddr_in *to, uint8_t* seq_number);
+int rudp_send(int sock_id, void *data, size_t data_size, int flags, struct sockaddr_in *to, uint16_t* seq_number);
 
 /* 
  * @brief Receives data from peer.
